refactor: split rrweb-snapshot utilities
